0.5.21 - Added settings update notification
